playlist refactor

+++ internal/infra/mongodb/mongo_playlist_repository.go
@@ -7,6 +7,7 @@ import (
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
 type MongoPlaylistRepository struct {
@@ -28,22 +29,30 @@ func (r *MongoPlaylistRepository) Create(ctx context.Context, playlist *models.P
 	return result.InsertedID.(primitive.ObjectID).Hex(), nil
 }
 
-func (r *MongoPlaylistRepository) FindByID(ctx context.Context, id string) (*models.Playlist, error) {
+func (r *MongoPlaylistRepository) FindByID(ctx context.Context, id string, ownerID string) (*models.Playlist, error) {
 	playlist := &MongoPlaylist{}
 	idPrimitive, err := primitive.ObjectIDFromHex(id)
 	if err != nil {
 		return nil, err
 	}
-	err = r.playlistsColl.FindOne(ctx, bson.M{"_id": idPrimitive}).Decode(playlist)
+	ownerIDPrimitive, err := primitive.ObjectIDFromHex(ownerID)
+	if err != nil {
+		return nil, err
+	}
+	err = r.playlistsColl.FindOne(ctx, bson.M{"_id": idPrimitive, "owner_id": ownerIDPrimitive}).Decode(playlist)
 	if err != nil {
 		return nil, err
 	}
 	return playlist.ToModel(), nil
 }
 
-func (r *MongoPlaylistRepository) FindByUserID(ctx context.Context, userID string) ([]*models.Playlist, error) {
+func (r *MongoPlaylistRepository) List(ctx context.Context, ownerID string, page int, limit int) ([]*models.Playlist, error) {
 	playlists := []*MongoPlaylist{}
-	cursor, err := r.playlistsColl.Find(ctx, bson.M{"user_id": userID})
+	ownerIDPrimitive, err := primitive.ObjectIDFromHex(ownerID)
+	if err != nil {
+		return nil, err
+	}
+	cursor, err := r.playlistsColl.Find(ctx, bson.M{"owner_id": ownerIDPrimitive}, options.Find().SetSkip(int64(page*limit)).SetLimit(int64(limit)))
 	if err != nil {
 		return nil, err
 	}
@@ -65,15 +74,15 @@ func (r *MongoPlaylistRepository) FindByUserID(ctx context.Context, userID strin
 func (r *MongoPlaylistRepository) Update(ctx context.Context, playlist *models.Playlist) error {
 	mongoPlaylist := &MongoPlaylist{}
 	mongoPlaylist.ByModel(playlist)
-	_, err := r.playlistsColl.UpdateOne(ctx, bson.M{"_id": playlist.ID}, bson.M{"$set": mongoPlaylist})
+	_, err := r.playlistsColl.UpdateOne(ctx, bson.M{"_id": playlist.ID, "owner_id": playlist.OwnerID}, bson.M{"$set": mongoPlaylist})
 	return err
 }
 
-func (r *MongoPlaylistRepository) Delete(ctx context.Context, id string) error {
+func (r *MongoPlaylistRepository) Delete(ctx context.Context, id string, ownerID string) error {
 	idPrimitive, err := primitive.ObjectIDFromHex(id)
 	if err != nil {
 		return err
 	}
-	_, err = r.playlistsColl.DeleteOne(ctx, bson.M{"_id": idPrimitive})
+	_, err = r.playlistsColl.DeleteOne(ctx, bson.M{"_id": idPrimitive, "owner_id": ownerID})
 	return err
 }
